# api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: ShareSlices API
  summary: ShareSlices account and Artifact HTTP API contract.
  description: >
    This OpenAPI document describes the current ShareSlices HTTP API contract
    for account entry and the version 0.0.1 Artifact upload, Preview, Publish,
    and Viewer flow, including current-browser-session sign out. Password
    recovery, email verification, additional account sign-in methods, link
    rotation, private sharing, and administration are outside this contract.

    The 429 responses are reserved for a future rate limiting change and are
    not implemented; contract tests exclude them until that change ships.
  version: 0.0.1
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /
    description: Same-origin API server.
...
security: []
paths:
...
  /api/versions/{versionId}/content/:
    parameters:
      - $ref: "#/components/parameters/VersionId"
    get:
      tags:
        - preview
      operationId: previewVersionEntry
      summary: Preview a ready Version's root entry file.
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Ready root entry streamed with caching disabled.
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            text/html:
              schema:
                type: string
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "404":
          $ref: "#/components/responses/VersionNotFound"
        "409":
          $ref: "#/components/responses/VersionNotReady"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/versions/{versionId}/export:
    parameters:
      - $ref: "#/components/parameters/VersionId"
    get:
      tags:
        - artifacts
      operationId: exportArtifactVersion
      summary: Download an owned ready Version as a ZIP.
      security:
        - sessionCookie: []
        - sessionBearer: []
      parameters:
        - $ref: "#/components/parameters/CliVersionWhenBearer"
        - $ref: "#/components/parameters/CliOperatingSystemWhenBearer"
        - name: artifactId
          in: query
          required: false
          description: Optional owning Artifact constraint used by CLI Export.
          schema:
            type: string
            minLength: 1
      responses:
        "200":
          description: ZIP assembled from committed Version assets.
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
            Content-Disposition:
              description: Safe attachment filename derived from the Artifact name.
              schema:
                type: string
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/zip:
              schema:
                type: string
                format: binary
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "404":
          $ref: "#/components/responses/VersionNotFound"
        "426":
          $ref: "#/components/responses/CliUpgradeRequired"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/versions/{versionId}/thumbnail:
    parameters:
      - $ref: "#/components/parameters/VersionId"
    get:
      tags:
        - artifacts
      operationId: getArtifactVersionThumbnail
      summary: Read the immutable thumbnail for an owned ready Version.
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Private immutable WebP thumbnail.
          headers:
            Cache-Control:
              description: Private immutable browser cache policy.
              schema:
                type: string
                const: private, max-age=31536000, immutable
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            image/webp:
              schema:
                type: string
                format: binary
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "404":
          $ref: "#/components/responses/ThumbnailNotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/artifacts/{artifactId}/publications:
    parameters:
      - $ref: "#/components/parameters/ArtifactId"
    post:
      tags:
        - publications
      operationId: createPublication
      summary: Publish a ready Version atomically.
      security:
        - sessionCookie: []
        - sessionBearer: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/CliVersionWhenBearer"
        - $ref: "#/components/parameters/CliOperatingSystemWhenBearer"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePublicationRequest"
      responses:
        "201":
          description: Publication created or original idempotent result returned.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicationResponse"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "404":
          $ref: "#/components/responses/ArtifactNotFound"
        "409":
          $ref: "#/components/responses/OperationConflict"
        "426":
          $ref: "#/components/responses/CliUpgradeRequired"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/artifacts/{artifactId}/publications/{publicationId}:
    parameters:
      - $ref: "#/components/parameters/ArtifactId"
      - $ref: "#/components/parameters/PublicationId"
    patch:
      tags:
        - publications
      operationId: updatePublication
      summary: Change the accessible Publication expiration without publishing again.
      security:
        - sessionCookie: []
        - sessionBearer: []
      parameters:
        - $ref: "#/components/parameters/CliVersionWhenBearer"
        - $ref: "#/components/parameters/CliOperatingSystemWhenBearer"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePublicationRequest"
      responses:
        "200":
          description: Publication expiration updated; Version and Share link are unchanged.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicationResponse"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "404":
          $ref: "#/components/responses/ArtifactNotFound"
        "409":
          $ref: "#/components/responses/OperationConflict"
        "426":
          $ref: "#/components/responses/CliUpgradeRequired"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      tags:
        - publications
      operationId: deletePublication
      summary: Unpublish while retaining the active Share link.
      security:
        - sessionCookie: []
        - sessionBearer: []
      parameters:
        - $ref: "#/components/parameters/CliVersionWhenBearer"
        - $ref: "#/components/parameters/CliOperatingSystemWhenBearer"
      responses:
        "204":
          description: Artifact is unpublished; repeated deletes have the same effect.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "404":
          $ref: "#/components/responses/ArtifactNotFound"
        "426":
          $ref: "#/components/responses/CliUpgradeRequired"
        "500":
          $ref: "#/components/responses/InternalError"
  /a/{shareSlug}/:
    parameters:
      - $ref: "#/components/parameters/ShareSlug"
    get:
      tags:
        - viewer
      operationId: getSharedArtifactEntry
      summary: View published content or the known Share-link state.
      parameters:
        - name: contentMode
          in: query
          required: false
          description: Reserved player-frame mode that returns the published manifest entry instead of the trusted Viewer shell.
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Trusted Viewer player, published entry content in content mode, or an unpublished state page.
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            text/html:
              schema:
                type: string
        "410":
          description: Known Share link is expired or retired.
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            text/html:
              schema:
                type: string
        "404":
          description: Share link is unknown.
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            text/html:
              schema:
                type: string
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: shareslices_session
      description: Browser session cookie established by successful email/password login.
    sessionBearer:
      type: http
      scheme: bearer
      description: Opaque independently revocable CLI Session credential.
  parameters:
...
    CliVersionWhenBearer:
      name: ShareSlices-CLI-Version
      in: header
      required: false
      schema:
        type: string
        pattern: "^[0-9]+\\.[0-9]+\\.[0-9]+$"
      description: Required when the selected security scheme is sessionBearer; not persisted.
    CliOperatingSystemWhenBearer:
      name: ShareSlices-CLI-OS
      in: header
      required: false
      schema:
        type: string
        enum: [linux, macos, windows]
      description: Required when the selected security scheme is sessionBearer; not persisted.
...
    ArtifactId:
      name: artifactId
      in: path
      required: true
      description: Stable Artifact ID.
      schema:
        type: string
        minLength: 1
        maxLength: 128
...
    VersionId:
      name: versionId
      in: path
      required: true
      description: Stable immutable Version ID.
      schema:
        type: string
        minLength: 1
        maxLength: 128
    PublicationId:
      name: publicationId
      in: path
      required: true
      description: Stable Publication ID.
      schema:
        type: string
        minLength: 1
        maxLength: 128
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Caller-generated key scoped to the signed-in owner and operation.
      schema:
        type: string
        minLength: 1
        maxLength: 255
...
  headers:
    RequestId:
      description: Stable request identifier for logs and support.
      schema:
        type: string
        minLength: 1
...
    NoStore:
      description: Prevents storage of sensitive or state-dependent responses.
      required: true
      schema:
        type: string
        enum:
          - no-store
...
  responses:
...
    CliUpgradeRequired:
      description: The CLI is missing compatible metadata or must be upgraded.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            upgrade:
              value:
                error:
                  code: cli_upgrade_required
                  message: Update ShareSlices CLI to continue.
                  action: Update ShareSlices CLI to 0.1.0 or newer.
                  requestId: req_000000000000000000000000
                  details:
                    currentVersion: 0.0.9
                    minimumVersion: 0.1.0
                    operatingSystem: macos
                    supportedOperatingSystems: [linux, macos, windows]
...
    Unauthenticated:
      description: Signed-in state is missing or invalid.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            unauthenticated:
              value:
                error:
                  code: unauthenticated
                  message: Sign in to continue.
                  requestId: req_000000000000000000000000
...
    VersionNotFound:
      description: Version is unknown or is not owned by the current user.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            notFound:
              value:
                error:
                  code: version_not_found
                  message: Version not found.
                  requestId: req_000000000000000000000000
...
    ThumbnailNotFound:
      description: Thumbnail is pending, absent, failed, or not owned by the current user.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            notFound:
              value:
                error:
                  code: thumbnail_not_found
                  message: Thumbnail not found.
                  requestId: req_000000000000000000000000
    VersionNotReady:
      description: Version exists but has not reached the ready state.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            notReady:
              value:
                error:
                  code: version_not_ready
                  message: Version is not ready.
                  requestId: req_000000000000000000000000
...
    InternalError:
      description: The API encountered an unexpected failure.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            internalError:
              value:
                error:
                  code: internal_error
                  message: Internal server error.
                  requestId: req_000000000000000000000000
...
  schemas:
...
    CreatePublicationRequest:
      type: object
      additionalProperties: false
      required:
        - versionId
        - expiration
        - link
      properties:
        versionId:
          type: string
          minLength: 1
          maxLength: 128
        expiration:
          $ref: "#/components/schemas/PublicationExpirationPolicy"
        link:
          $ref: "#/components/schemas/PublicationLinkChoice"
...